requirements.txt
async-timeout==4.0.3
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SSE==1.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
python-engineio==4.9.1
python-socketio==5.11.2
redis==5.0.4
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
tempmail-python==2.3.3
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
